Move LexSmartTable into its own files
LexSmartTable is the SmartTable-based reference encoding kept around
for benchmarking. Putting it alongside the dedicated Lex propagator in
lex.{hh,cc} muddied both: lex.cc had to drag in smart_table.hh, and
lex.hh ran together two unrelated constraints. Split it out into
gcs/constraints/lex_smart_table.{hh,cc}, register the new source in
CMakeLists.txt, and update the smart_table_lex example to include the
new header. The header points users at LexGreaterThan (and the rest of
the lex.hh family) for normal use.
